+++ src/main/java/ru/savini/fb/repo/CategoryRepo.java
@@ -5,6 +5,6 @@ import ru.savini.fb.domain.entity.Category;
 
 import java.util.List;
 
-public interface CategoryRepo extends JpaRepository<Category, Long> {
+public interface CategoryRepo extends JpaRepository<Category, Integer> {
     List<Category> findByNameStartsWithIgnoreCase(String name);
 }
